refactor: extract agent_executor, graph, and templating from OrchestratorBase
Decompose the 1942-line OrchestratorBase into three focused modules:
- agent_executor.py: AgentExecutor class with callback-based execute/validate
- graph.py: build_execution_graph, get_ready_prompts, evaluate_condition
- templating.py: resolve_variables, resolve_prompt_variables, resolve_batch_name

OrchestratorBase retains thin delegate methods for backward compatibility.
Tests migrated to new test files with enhanced coverage (1793 pass, 0 fail).
