+++ tests/test_excel_orchestrator.py
@@ -1114,104 +1114,6 @@ class TestExcelOrchestratorParallelExecution:
         assert all(r["status"] == "success" for r in results)
 
 
-class TestExcelOrchestratorBatchVariableResolution:
-    """Tests for variable resolution in batch mode."""
-
-    def test_resolve_variables_basic(self, temp_workbook_with_batch_data, mock_ffmistralsmall):
-        """Test basic variable resolution."""
-        from src.orchestrator.excel_orchestrator import ExcelOrchestrator
-
-        orchestrator = ExcelOrchestrator(temp_workbook_with_batch_data, mock_ffmistralsmall)
-
-        data_row = {"region": "north", "product": "widget_a", "price": 10}
-
-        result = orchestrator._resolve_variables(
-            "Analyze {{region}} region, product {{product}}.", data_row
-        )
-
-        assert result == "Analyze north region, product widget_a."
-
-    def test_resolve_variables_multiple_occurrences(
-        self, temp_workbook_with_batch_data, mock_ffmistralsmall
-    ):
-        """Test variable resolution with multiple occurrences."""
-        from src.orchestrator.excel_orchestrator import ExcelOrchestrator
-
-        orchestrator = ExcelOrchestrator(temp_workbook_with_batch_data, mock_ffmistralsmall)
-
-        data_row = {"region": "east", "price": 20}
-
-        result = orchestrator._resolve_variables(
-            "{{region}} has price {{price}} and tax on {{price}}.", data_row
-        )
-
-        assert result == "east has price 20 and tax on 20."
-
-    def test_resolve_variables_missing_keeps_placeholder(
-        self, temp_workbook_with_batch_data, mock_ffmistralsmall
-    ):
-        """Test that missing variables keep placeholder."""
-        from src.orchestrator.excel_orchestrator import ExcelOrchestrator
-
-        orchestrator = ExcelOrchestrator(temp_workbook_with_batch_data, mock_ffmistralsmall)
-
-        data_row = {"region": "north"}
-
-        result = orchestrator._resolve_variables(
-            "Region: {{region}}, Missing: {{unknown}}", data_row
-        )
-
-        assert result == "Region: north, Missing: {{unknown}}"
-
-    def test_resolve_prompt_variables(self, temp_workbook_with_batch_data, mock_ffmistralsmall):
-        """Test resolving variables in prompt dict."""
-        from src.orchestrator.excel_orchestrator import ExcelOrchestrator
-
-        orchestrator = ExcelOrchestrator(temp_workbook_with_batch_data, mock_ffmistralsmall)
-
-        prompt = {
-            "sequence": 1,
-            "prompt_name": "test",
-            "prompt": "Price is {{price}}, quantity is {{quantity}}.",
-        }
-        data_row = {"price": 10, "quantity": 100}
-
-        result = orchestrator._resolve_prompt_variables(prompt, data_row)
-
-        assert result["prompt"] == "Price is 10, quantity is 100."
-        assert result["sequence"] == 1
-
-    def test_resolve_batch_name_with_template(
-        self, temp_workbook_with_batch_data, mock_ffmistralsmall
-    ):
-        """Test batch name resolution with template."""
-        from src.orchestrator.excel_orchestrator import ExcelOrchestrator
-
-        orchestrator = ExcelOrchestrator(temp_workbook_with_batch_data, mock_ffmistralsmall)
-
-        data_row = {
-            "region": "north",
-            "product": "widget_a",
-            "batch_name": "{{region}}_{{product}}",
-        }
-
-        result = orchestrator._resolve_batch_name(data_row, 1)
-
-        assert result == "north_widget_a"
-
-    def test_resolve_batch_name_default(self, temp_workbook_with_batch_data, mock_ffmistralsmall):
-        """Test batch name resolution with default."""
-        from src.orchestrator.excel_orchestrator import ExcelOrchestrator
-
-        orchestrator = ExcelOrchestrator(temp_workbook_with_batch_data, mock_ffmistralsmall)
-
-        data_row = {"region": "north", "product": "widget_a"}
-
-        result = orchestrator._resolve_batch_name(data_row, 5)
-
-        assert result == "batch_5"
-
-
 class TestExcelOrchestratorBatchExecution:
     """Tests for batch execution."""
