+++ tests/test_manifest_comprehensive.py
@@ -339,152 +339,6 @@ class TestManifestOrchestratorDocuments:
         assert orchestrator._parse_bool_override("invalid") is None
 
 
-class TestManifestOrchestratorVariables:
-    """Tests for ManifestOrchestrator variable resolution."""
-
-    def test_resolve_variables_basic(self, tmp_path, mock_ffmistralsmall):
-        """Test _resolve_variables replaces placeholders."""
-        from src.orchestrator.manifest import ManifestOrchestrator
-
-        orchestrator = ManifestOrchestrator(
-            manifest_dir=str(tmp_path),
-            client=mock_ffmistralsmall,
-        )
-
-        text = "Region: {{region}}, Product: {{product}}"
-        data_row = {"region": "North", "product": "Widget"}
-
-        result = orchestrator._resolve_variables(text, data_row)
-
-        assert result == "Region: North, Product: Widget"
-
-    def test_resolve_variables_missing(self, tmp_path, mock_ffmistralsmall):
-        """Test _resolve_variables keeps placeholder for missing variable."""
-        from src.orchestrator.manifest import ManifestOrchestrator
-
-        orchestrator = ManifestOrchestrator(
-            manifest_dir=str(tmp_path),
-            client=mock_ffmistralsmall,
-        )
-
-        text = "Region: {{region}}, Missing: {{unknown}}"
-        data_row = {"region": "North"}
-
-        result = orchestrator._resolve_variables(text, data_row)
-
-        assert result == "Region: North, Missing: {{unknown}}"
-
-    def test_resolve_variables_empty_text(self, tmp_path, mock_ffmistralsmall):
-        """Test _resolve_variables with empty text."""
-        from src.orchestrator.manifest import ManifestOrchestrator
-
-        orchestrator = ManifestOrchestrator(
-            manifest_dir=str(tmp_path),
-            client=mock_ffmistralsmall,
-        )
-
-        result = orchestrator._resolve_variables("", {"region": "North"})
-
-        assert result == ""
-
-    def test_resolve_variables_none_value(self, tmp_path, mock_ffmistralsmall):
-        """Test _resolve_variables with None value in data."""
-        from src.orchestrator.manifest import ManifestOrchestrator
-
-        orchestrator = ManifestOrchestrator(
-            manifest_dir=str(tmp_path),
-            client=mock_ffmistralsmall,
-        )
-
-        text = "Value: {{value}}"
-        data_row = {"value": None}
-
-        result = orchestrator._resolve_variables(text, data_row)
-
-        assert result == "Value: {{value}}"
-
-    def test_resolve_prompt_variables(self, tmp_path, mock_ffmistralsmall):
-        """Test _resolve_prompt_variables resolves all fields."""
-        from src.orchestrator.manifest import ManifestOrchestrator
-
-        orchestrator = ManifestOrchestrator(
-            manifest_dir=str(tmp_path),
-            client=mock_ffmistralsmall,
-        )
-
-        prompt = {
-            "prompt": "Hello {{name}}",
-            "prompt_name": "greeting_{{region}}",
-            "sequence": 1,
-        }
-        data_row = {"name": "World", "region": "North"}
-
-        result = orchestrator._resolve_prompt_variables(prompt, data_row)
-
-        assert result["prompt"] == "Hello World"
-        assert result["prompt_name"] == "greeting_North"
-        assert result["sequence"] == 1
-
-    def test_resolve_batch_name_custom(self, tmp_path, mock_ffmistralsmall):
-        """Test _resolve_batch_name with custom template."""
-        from src.orchestrator.manifest import ManifestOrchestrator
-
-        orchestrator = ManifestOrchestrator(
-            manifest_dir=str(tmp_path),
-            client=mock_ffmistralsmall,
-        )
-
-        data_row = {"batch_name": "{{region}}_{{product}}", "region": "north", "product": "widget"}
-        result = orchestrator._resolve_batch_name(data_row, 1)
-
-        assert result == "north_widget"
-
-    def test_resolve_batch_name_sanitizes_special_chars(self, tmp_path, mock_ffmistralsmall):
-        """Test _resolve_batch_name sanitizes special characters."""
-        from src.orchestrator.manifest import ManifestOrchestrator
-
-        orchestrator = ManifestOrchestrator(
-            manifest_dir=str(tmp_path),
-            client=mock_ffmistralsmall,
-        )
-
-        data_row = {"batch_name": "test @#$ batch!!!", "region": "north"}
-        result = orchestrator._resolve_batch_name(data_row, 1)
-
-        assert "@" not in result
-        assert "#" not in result
-        assert "!" not in result
-
-    def test_resolve_batch_name_default(self, tmp_path, mock_ffmistralsmall):
-        """Test _resolve_batch_name uses default when no batch_name field."""
-        from src.orchestrator.manifest import ManifestOrchestrator
-
-        orchestrator = ManifestOrchestrator(
-            manifest_dir=str(tmp_path),
-            client=mock_ffmistralsmall,
-        )
-
-        data_row = {"region": "north"}
-        result = orchestrator._resolve_batch_name(data_row, 5)
-
-        assert result == "batch_5"
-
-    def test_resolve_batch_name_truncates_long_names(self, tmp_path, mock_ffmistralsmall):
-        """Test _resolve_batch_name truncates names longer than 50 chars."""
-        from src.orchestrator.manifest import ManifestOrchestrator
-
-        orchestrator = ManifestOrchestrator(
-            manifest_dir=str(tmp_path),
-            client=mock_ffmistralsmall,
-        )
-
-        long_name = "a" * 100
-        data_row = {"batch_name": long_name}
-        result = orchestrator._resolve_batch_name(data_row, 1)
-
-        assert len(result) == 50
-
-
 class TestManifestOrchestratorConditionEvaluation:
     """Tests for ManifestOrchestrator condition evaluation."""
