+++ tests/test_orchestrator_base.py
@@ -5,7 +5,6 @@
 """Tests for OrchestratorBase fixes and improvements.
 
 Tests cover:
-- Dependency cycle detection (#6)
 - Retry backoff delay (#5)
 - Unified execution path / client clone reuse (#1, #7)
 - Consistent logging (#8)
@@ -16,8 +15,7 @@ Tests cover:
 - Agent result recording (332, 347)
 - Validation edge cases (364, 373-374)
 - Reference injection with semantic query (422-439, 452)
-- Agent mode execution paths (612-616, 658, 660, 686-688, 894)
-- Agent response validation (742-752, 820-822)
+- Agent mode fallback in _execute_with_retry (894)
 """
 
 import json
@@ -28,110 +26,6 @@ from unittest.mock import MagicMock, patch
 import pytest
 
 from src.orchestrator.base.orchestrator_base import OrchestratorBase
-
-
-class TestDependencyCycleDetection:
-    """Tests for dependency cycle detection in _build_execution_graph."""
-
-    def _make_orchestrator(self, prompts, mock_ffmistralsmall):
-        """Helper to create an orchestrator with preset prompts."""
-        from src.orchestrator.excel_orchestrator import ExcelOrchestrator
-
-        orchestrator = ExcelOrchestrator.__new__(ExcelOrchestrator)
-        # Minimal init for testing _build_execution_graph
-        orchestrator.client = mock_ffmistralsmall
-        orchestrator.prompts = prompts
-        return orchestrator
-
-    def test_cycle_detection_simple(self, temp_workbook, mock_ffmistralsmall):
-        """Test that a simple A->B->A cycle is detected."""
-        from src.orchestrator.excel_orchestrator import ExcelOrchestrator
-
-        orchestrator = ExcelOrchestrator(temp_workbook, mock_ffmistralsmall)
-        orchestrator.prompts = [
-            {"sequence": 1, "prompt_name": "a", "prompt": "Hello", "history": ["b"]},
-            {"sequence": 2, "prompt_name": "b", "prompt": "World", "history": ["a"]},
-        ]
-
-        with pytest.raises(ValueError, match="Dependency cycle detected"):
-            orchestrator._build_execution_graph()
-
-    def test_cycle_detection_three_node(self, temp_workbook, mock_ffmistralsmall):
-        """Test that a A->B->C->A cycle is detected."""
-        from src.orchestrator.excel_orchestrator import ExcelOrchestrator
-
-        orchestrator = ExcelOrchestrator(temp_workbook, mock_ffmistralsmall)
-        orchestrator.prompts = [
-            {"sequence": 1, "prompt_name": "a", "prompt": "A", "history": ["c"]},
-            {"sequence": 2, "prompt_name": "b", "prompt": "B", "history": ["a"]},
-            {"sequence": 3, "prompt_name": "c", "prompt": "C", "history": ["b"]},
-        ]
-
-        with pytest.raises(ValueError, match="Dependency cycle detected"):
-            orchestrator._build_execution_graph()
-
-    def test_self_referencing_cycle(self, temp_workbook, mock_ffmistralsmall):
-        """Test that a self-referencing dependency is detected."""
-        from src.orchestrator.excel_orchestrator import ExcelOrchestrator
-
-        orchestrator = ExcelOrchestrator(temp_workbook, mock_ffmistralsmall)
-        orchestrator.prompts = [
-            {"sequence": 1, "prompt_name": "a", "prompt": "Hello", "history": ["a"]},
-        ]
-
-        with pytest.raises(ValueError, match="Dependency cycle detected"):
-            orchestrator._build_execution_graph()
-
-    def test_no_cycle_linear_chain(self, temp_workbook, mock_ffmistralsmall):
-        """Test that a valid linear chain does not raise."""
-        from src.orchestrator.excel_orchestrator import ExcelOrchestrator
-
-        orchestrator = ExcelOrchestrator(temp_workbook, mock_ffmistralsmall)
-        orchestrator.prompts = [
-            {"sequence": 1, "prompt_name": "a", "prompt": "A", "history": None},
-            {"sequence": 2, "prompt_name": "b", "prompt": "B", "history": ["a"]},
-            {"sequence": 3, "prompt_name": "c", "prompt": "C", "history": ["b"]},
-        ]
-
-        nodes = orchestrator._build_execution_graph()
-
-        assert nodes[1].level == 0
-        assert nodes[2].level == 1
-        assert nodes[3].level == 2
-
-    def test_no_cycle_diamond(self, temp_workbook, mock_ffmistralsmall):
-        """Test that a valid diamond dependency pattern does not raise."""
-        from src.orchestrator.excel_orchestrator import ExcelOrchestrator
-
-        orchestrator = ExcelOrchestrator(temp_workbook, mock_ffmistralsmall)
-        orchestrator.prompts = [
-            {"sequence": 1, "prompt_name": "a", "prompt": "A", "history": None},
-            {"sequence": 2, "prompt_name": "b", "prompt": "B", "history": ["a"]},
-            {"sequence": 3, "prompt_name": "c", "prompt": "C", "history": ["a"]},
-            {"sequence": 4, "prompt_name": "d", "prompt": "D", "history": ["b", "c"]},
-        ]
-
-        nodes = orchestrator._build_execution_graph()
-
-        assert nodes[1].level == 0
-        assert nodes[2].level == 1
-        assert nodes[3].level == 1
-        assert nodes[4].level == 2
-
-    def test_no_cycle_independent(self, temp_workbook, mock_ffmistralsmall):
-        """Test that independent prompts all get level 0."""
-        from src.orchestrator.excel_orchestrator import ExcelOrchestrator
-
-        orchestrator = ExcelOrchestrator(temp_workbook, mock_ffmistralsmall)
-        orchestrator.prompts = [
-            {"sequence": 1, "prompt_name": "a", "prompt": "A", "history": None},
-            {"sequence": 2, "prompt_name": "b", "prompt": "B", "history": None},
-            {"sequence": 3, "prompt_name": "c", "prompt": "C", "history": None},
-        ]
-
-        nodes = orchestrator._build_execution_graph()
-
-        assert all(node.level == 0 for node in nodes.values())
 
 
 class TestRetryBackoff:
@@ -813,30 +707,6 @@ class TestInitTools:
         assert "custom_tool" in registered_names
 
 
-class TestRecordAgentResult:
-    """Tests for _record_agent_result_in_shared_history covering lines 332, 347."""
-
-    def test_none_response_returns_early(self, mock_ffmistralsmall):
-        """Test that None response returns immediately (line 332)."""
-        orch = _make_base(mock_ffmistralsmall)
-        orch.config = {"model": "test"}
-        initial_len = len(orch.shared_prompt_attr_history)
-        orch._record_agent_result_in_shared_history({"prompt": "hi"}, None)
-        assert len(orch.shared_prompt_attr_history) == initial_len
-
-    def test_no_history_lock(self, mock_ffmistralsmall):
-        """Test recording when history_lock is None (line 347)."""
-        orch = _make_base(mock_ffmistralsmall)
-        orch.history_lock = None
-        orch.config = {"model": "test"}
-        orch._record_agent_result_in_shared_history(
-            {"prompt": "hi", "prompt_name": "test", "history": None},
-            "response_text",
-        )
-        assert len(orch.shared_prompt_attr_history) == 1
-        assert orch.shared_prompt_attr_history[0]["response"] == "response_text"
-
-
 class TestValidateEdgeCases:
     """Tests for _validate covering lines 364, 373-374."""
 
@@ -955,280 +825,6 @@ class TestInjectReferences:
         assert result == "original prompt"
 
 
-class TestExecuteAgentMode:
-    """Tests for _execute_agent_mode covering lines 612-616, 658, 660, 686-688."""
-
-    def test_openai_assistant_fallback(self, mock_ffmistralsmall):
-        """Test FFOpenAIAssistant fallback to single-shot (lines 612-616)."""
-        orch = _make_base(mock_ffmistralsmall)
-        orch.config = {"model": "gpt-4"}
-        orch.tool_registry = MagicMock()
-
-        mock_client = MagicMock()
-        mock_client.__class__.__name__ = "FFOpenAIAssistant"
-        mock_ffai = MagicMock()
-        mock_ffai.client = mock_client
-
-        mock_builder = MagicMock()
-        mock_builder.build_dict.return_value = {"status": "success"}
-
-        prompt = {
-            "sequence": 1,
-            "prompt_name": "test",
-            "prompt": "Hello",
-            "history": None,
-            "agent_mode": True,
-            "tools": ["calculate"],
-        }
-
-        result = orch._execute_agent_mode(prompt, mock_ffai, mock_builder, "Sequence 1")
-        assert result is None
-
-    def test_agent_result_failed(self, mock_ffmistralsmall):
-        """Test agent result with status failed (line 658)."""
-        from src.agent.agent_result import AgentResult
-
-        orch = _make_base(mock_ffmistralsmall)
-        orch.config = {"model": "test", "temperature": 0.7, "max_tokens": 4096}
-
-        mock_agent_result = AgentResult(
-            response="",
-            total_rounds=1,
-            total_llm_calls=1,
-            status="failed",
-        )
-
-        mock_tool_reg = MagicMock()
-        orch.tool_registry = mock_tool_reg
-
-        mock_ffai = MagicMock()
-        mock_ffai.client = mock_ffmistralsmall
-        mock_ffai._build_prompt.return_value = ("resolved prompt", None)
-        mock_ffai.last_resolved_prompt = "resolved prompt"
-
-        mock_builder = MagicMock()
-        mock_builder.build_dict.return_value = {"status": "failed"}
-
-        mock_config = MagicMock()
-        mock_config.agent.max_tool_rounds = 5
-        mock_config.agent.tool_timeout = 30.0
-        mock_config.agent.continue_on_tool_error = True
-        mock_config.agent.validation.enabled = False
-
-        with patch("src.orchestrator.base.orchestrator_base.get_config", return_value=mock_config):
-            with patch("src.agent.agent_loop.AgentLoop") as mock_loop_cls:
-                mock_loop = MagicMock()
-                mock_loop.execute.return_value = mock_agent_result
-                mock_loop_cls.return_value = mock_loop
-
-                orch._execute_agent_mode(
-                    {
-                        "sequence": 1,
-                        "prompt_name": "test",
-                        "prompt": "Hello",
-                        "history": None,
-                        "agent_mode": True,
-                        "tools": ["calculate"],
-                    },
-                    mock_ffai,
-                    mock_builder,
-                    "Sequence 1",
-                )
-
-        mock_builder.with_agent_result.assert_called_once_with(mock_agent_result)
-
-    def test_agent_result_max_rounds_exceeded(self, mock_ffmistralsmall):
-        """Test agent result with status max_rounds_exceeded (line 660)."""
-        from src.agent.agent_result import AgentResult
-
-        orch = _make_base(mock_ffmistralsmall)
-        orch.config = {"model": "test", "temperature": 0.7, "max_tokens": 4096}
-
-        mock_agent_result = AgentResult(
-            response="partial",
-            total_rounds=5,
-            total_llm_calls=5,
-            status="max_rounds_exceeded",
-        )
-
-        mock_tool_reg = MagicMock()
-        orch.tool_registry = mock_tool_reg
-
-        mock_ffai = MagicMock()
-        mock_ffai.client = mock_ffmistralsmall
-        mock_ffai._build_prompt.return_value = ("resolved prompt", None)
-        mock_ffai.last_resolved_prompt = "resolved prompt"
-
-        mock_builder = MagicMock()
-        mock_builder.build_dict.return_value = {"status": "max_rounds_exceeded"}
-
-        mock_config = MagicMock()
-        mock_config.agent.max_tool_rounds = 5
-        mock_config.agent.tool_timeout = 30.0
-        mock_config.agent.continue_on_tool_error = True
-        mock_config.agent.validation.enabled = False
-
-        with patch("src.orchestrator.base.orchestrator_base.get_config", return_value=mock_config):
-            with patch("src.agent.agent_loop.AgentLoop") as mock_loop_cls:
-                mock_loop = MagicMock()
-                mock_loop.execute.return_value = mock_agent_result
-                mock_loop_cls.return_value = mock_loop
-
-                orch._execute_agent_mode(
-                    {
-                        "sequence": 1,
-                        "prompt_name": "test",
-                        "prompt": "Hello",
-                        "history": None,
-                        "agent_mode": True,
-                        "tools": ["calculate"],
-                    },
-                    mock_ffai,
-                    mock_builder,
-                    "Sequence 1",
-                )
-
-        mock_builder.with_agent_result.assert_called_once_with(mock_agent_result)
-
-    def test_agent_mode_exception(self, mock_ffmistralsmall):
-        """Test exception in agent execution (lines 686-688)."""
-        orch = _make_base(mock_ffmistralsmall)
-        orch.config = {"model": "test", "temperature": 0.7, "max_tokens": 4096}
-        orch.tool_registry = MagicMock()
-
-        mock_ffai = MagicMock()
-        mock_ffai.client = mock_ffmistralsmall
-        mock_ffai._build_prompt.return_value = ("resolved prompt", None)
-        mock_ffai.last_resolved_prompt = "resolved prompt"
-
-        mock_builder = MagicMock()
-        mock_builder.build_dict.return_value = {"status": "failed", "error": "boom"}
-
-        mock_config = MagicMock()
-        mock_config.agent.max_tool_rounds = 5
-        mock_config.agent.tool_timeout = 30.0
-        mock_config.agent.continue_on_tool_error = True
-
-        with patch("src.orchestrator.base.orchestrator_base.get_config", return_value=mock_config):
-            with patch("src.agent.agent_loop.AgentLoop") as mock_loop_cls:
-                mock_loop_cls.return_value = MagicMock()
-                mock_loop_cls.return_value.execute.side_effect = RuntimeError("agent exploded")
-
-                result = orch._execute_agent_mode(
-                    {
-                        "sequence": 1,
-                        "prompt_name": "test",
-                        "prompt": "Hello",
-                        "history": None,
-                        "agent_mode": True,
-                        "tools": ["calculate"],
-                    },
-                    mock_ffai,
-                    mock_builder,
-                    "Sequence 1",
-                )
-
-        mock_builder.with_error.assert_called_once()
-        assert "agent exploded" in mock_builder.with_error.call_args[0][0]
-
-
-class TestValidateAgentResponse:
-    """Tests for _validate_agent_response covering lines 742-752, 820-822."""
-
-    def test_validation_llm_fails_all_retries(self, mock_ffmistralsmall):
-        """Test when validation LLM call fails and retries are exhausted (lines 742-752)."""
-        from src.agent.agent_result import AgentResult
-
-        orch = _make_base(mock_ffmistralsmall)
-        orch.config = {"model": "test", "temperature": 0.7, "max_tokens": 4096}
-        orch.tool_registry = MagicMock()
-
-        mock_agent_result = AgentResult(response="original response", status="success")
-
-        mock_ffai_val = MagicMock()
-        mock_ffai_val.generate_response.side_effect = RuntimeError("LLM down")
-        orch._get_isolated_ffai = MagicMock(return_value=mock_ffai_val)
-
-        mock_builder = MagicMock()
-
-        mock_config = MagicMock()
-        mock_config.agent.max_tool_rounds = 5
-        mock_config.agent.tool_timeout = 30.0
-        mock_config.agent.continue_on_tool_error = True
-
-        with patch("src.orchestrator.base.orchestrator_base.get_config", return_value=mock_config):
-            with patch("src.agent.agent_loop.AgentLoop") as mock_loop_cls:
-                orch._validate_agent_response(
-                    prompt={"sequence": 1, "prompt_name": "test", "history": None},
-                    builder=mock_builder,
-                    agent_result=mock_agent_result,
-                    tool_names=["calculate"],
-                    validation_prompt="Must include numbers",
-                    max_val_retries=2,
-                    seq_label="Sequence 1",
-                    original_prompt="original prompt",
-                    max_rounds=5,
-                    tool_timeout=30.0,
-                    continue_on_tool_error=True,
-                )
-
-        mock_builder.with_validation_result.assert_called_once()
-        call_kwargs = mock_builder.with_validation_result.call_args[1]
-        assert call_kwargs["passed"] is None
-        assert "Validation call failed" in call_kwargs["critique"]
-
-    def test_validation_retry_execution_fails(self, mock_ffmistralsmall):
-        """Test when retry execution in validation loop fails (lines 820-822)."""
-        from src.agent.agent_result import AgentResult
-
-        orch = _make_base(mock_ffmistralsmall)
-        orch.config = {"model": "test", "temperature": 0.7, "max_tokens": 4096}
-        orch.tool_registry = MagicMock()
-
-        mock_agent_result = AgentResult(response="bad response", status="success")
-
-        call_count = [0]
-
-        def mock_val_generate(*args, **kwargs):
-            call_count[0] += 1
-            return "FAIL: Not good enough"
-
-        mock_ffai_val = MagicMock()
-        mock_ffai_val.generate_response = mock_val_generate
-        orch._get_isolated_ffai = MagicMock(return_value=mock_ffai_val)
-
-        mock_builder = MagicMock()
-
-        mock_config = MagicMock()
-        mock_config.agent.max_tool_rounds = 5
-        mock_config.agent.tool_timeout = 30.0
-        mock_config.agent.continue_on_tool_error = True
-
-        with patch("src.orchestrator.base.orchestrator_base.get_config", return_value=mock_config):
-            with patch("src.agent.agent_loop.AgentLoop") as mock_loop_cls:
-                mock_loop = MagicMock()
-                mock_loop.execute.side_effect = RuntimeError("retry failed")
-                mock_loop_cls.return_value = mock_loop
-
-                orch._validate_agent_response(
-                    prompt={"sequence": 1, "prompt_name": "test", "history": None},
-                    builder=mock_builder,
-                    agent_result=mock_agent_result,
-                    tool_names=["calculate"],
-                    validation_prompt="Must be perfect",
-                    max_val_retries=1,
-                    seq_label="Sequence 1",
-                    original_prompt="original prompt",
-                    max_rounds=5,
-                    tool_timeout=30.0,
-                    continue_on_tool_error=True,
-                )
-
-        last_call = mock_builder.with_validation_result.call_args[1]
-        assert last_call["passed"] is False
-        assert last_call["critique"] is not None
-
-
 class TestExecuteWithRetryAgentMode:
     """Tests for _execute_with_retry covering line 894."""
 
@@ -1277,122 +873,3 @@ class TestExecuteWithRetryAgentMode:
         assert result["response"] == "single shot response"
         warning_calls = [str(c) for c in mock_logger.warning.call_args_list]
         assert any("no tool registry" in msg for msg in warning_calls)
-
-
-class TestRowDocumentBinding:
-    """Tests for per-row document binding via _documents column (Phase 1)."""
-
-    def test_resolve_prompt_variables_with_documents(self, mock_ffmistralsmall):
-        """Test that _documents from data row are merged into references."""
-        orch = _make_base(mock_ffmistralsmall)
-
-        prompt = {
-            "sequence": 1,
-            "prompt_name": "evaluate",
-            "prompt": "Evaluate {{candidate}}",
-            "references": ["job_description"],
-        }
-        data_row = {"candidate": "Alice", "_documents": '["resume_alice"]'}
-
-        resolved = orch._resolve_prompt_variables(prompt, data_row)
-
-        assert resolved["prompt"] == "Evaluate Alice"
-        assert resolved["references"] == ["job_description", "resume_alice"]
-
-    def test_resolve_prompt_variables_additive_merge(self, mock_ffmistralsmall):
-        """Test additive merge: static references are preserved, _documents appended."""
-        orch = _make_base(mock_ffmistralsmall)
-
-        prompt = {
-            "sequence": 1,
-            "prompt_name": "eval",
-            "prompt": "Go",
-            "references": ["shared_doc", "rubric"],
-        }
-        data_row = {"_documents": '["resume_bob"]'}
-
-        resolved = orch._resolve_prompt_variables(prompt, data_row)
-
-        assert resolved["references"] == ["shared_doc", "rubric", "resume_bob"]
-
-    def test_resolve_prompt_variables_empty_static_refs(self, mock_ffmistralsmall):
-        """Test merge when prompt has no static references."""
-        orch = _make_base(mock_ffmistralsmall)
-
-        prompt = {
-            "sequence": 1,
-            "prompt_name": "eval",
-            "prompt": "Go",
-            "references": None,
-        }
-        data_row = {"_documents": '["resume_alice", "cover_letter_alice"]'}
-
-        resolved = orch._resolve_prompt_variables(prompt, data_row)
-
-        assert resolved["references"] == ["resume_alice", "cover_letter_alice"]
-
-    def test_resolve_prompt_variables_no_documents_column(self, mock_ffmistralsmall):
-        """Test that prompts without _documents are unchanged."""
-        orch = _make_base(mock_ffmistralsmall)
-
-        prompt = {
-            "sequence": 1,
-            "prompt_name": "eval",
-            "prompt": "Go",
-            "references": ["job_description"],
-        }
-        data_row = {"candidate": "Bob"}
-
-        resolved = orch._resolve_prompt_variables(prompt, data_row)
-
-        assert resolved["references"] == ["job_description"]
-
-    def test_resolve_prompt_variables_documents_empty_string(self, mock_ffmistralsmall):
-        """Test that empty _documents string is ignored."""
-        orch = _make_base(mock_ffmistralsmall)
-
-        prompt = {
-            "sequence": 1,
-            "prompt_name": "eval",
-            "prompt": "Go",
-            "references": ["jd"],
-        }
-        data_row = {"_documents": ""}
-
-        resolved = orch._resolve_prompt_variables(prompt, data_row)
-
-        assert resolved["references"] == ["jd"]
-
-    def test_resolve_prompt_variables_does_not_mutate_original(self, mock_ffmistralsmall):
-        """Test that _resolve_prompt_variables does not mutate the original prompt dict."""
-        orch = _make_base(mock_ffmistralsmall)
-
-        prompt = {
-            "sequence": 1,
-            "prompt_name": "eval",
-            "prompt": "Go",
-            "references": ["jd"],
-        }
-        original_refs = list(prompt["references"])
-        data_row = {"_documents": '["resume_alice"]'}
-
-        resolved = orch._resolve_prompt_variables(prompt, data_row)
-
-        assert prompt["references"] == original_refs
-        assert resolved["references"] == ["jd", "resume_alice"]
-
-    def test_resolve_prompt_variables_documents_json_list(self, mock_ffmistralsmall):
-        """Test _documents as a pre-parsed list (from data loading)."""
-        orch = _make_base(mock_ffmistralsmall)
-
-        prompt = {
-            "sequence": 1,
-            "prompt_name": "eval",
-            "prompt": "Go",
-            "references": ["jd"],
-        }
-        data_row = {"_documents": ["resume_alice", "cover_letter"]}
-
-        resolved = orch._resolve_prompt_variables(prompt, data_row)
-
-        assert resolved["references"] == ["jd", "resume_alice", "cover_letter"]
